pyproject.toml
[project]
name = "customer-service-agent"
version = "0.1.0"
description = "Trilingual customer service multi-agent system using Vertex AI ADK"
requires-python = ">=3.10,<3.13"
dependencies = [
    "google-adk>=1.18.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "google-cloud-storage>=2.0.0",
    "google-cloud-logging>=3.0.0",
    "opentelemetry-sdk>=1.20.0",
    "langfuse>=2.0.0",
    "openinference-instrumentation-google-adk>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

lint = [
    "ruff>=0.4.6",
    "mypy~=1.15.0",
    "codespell~=2.2.0",
]

[tool.ruff]
line-length = 80
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",
    "F",
    "W",
    "I",
    "C",
    "B",
    "UP",
    "RUF",
]
ignore = ["E501", "C901"]

[tool.ruff.lint.isort]
known-first-party = ["src", "backend"]

[tool.mypy]
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
check_untyped_defs = true
warn_redundant_casts = true
warn_unused_ignores = true
follow_imports = "silent"
ignore_missing_imports = true
explicit_package_bases = true
disable_error_code = ["misc", "no-untyped-call", "no-any-return"]
exclude = [".venv"]

[tool.codespell]
skip = ".venv,uv.lock,**/*.ipynb"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
pythonpath = "."
asyncio_default_fixture_loop_scope = "function"

[tool.hatch.build.targets.wheel]
packages = ["backend/src"]
